drivers: debug: Fix AXI timeout can't trigger EMMD bug
When AXI timeout occurs on cpux, if cpux continue accesses reg under s2
port, cpux will waiting for s2 reg and can't execute flush_cache_all().
Thus, the saved variables can't be stored in DDR.

Also, variable err_addr should be extended to 64bits to support ulc.

+++ drivers/marvell/debug/mmp-debug.c
@@ -48,8 +48,8 @@ static void __iomem *squ_base;
 
 static u32 timeout_write_addr, timeout_read_addr;
 static u32 finish_save_cpu_ctx;
-static u32 err_fsr;
-static u32 err_addr;
+static unsigned int  err_fsr;
+static unsigned long err_addr;
 
 static struct held_status recorded_helds;
